Move chunkify and format_file_size
