+++ perun/utils/log.py
@@ -734,6 +734,30 @@ def scan_formatting_string(
     return tokens
 
 
+def format_file_size(size: Optional[float]) -> str:
+    """Format file size in Bytes into a fixed-length output so that it can be easily printed.
+
+    If size is set to 'None' then the function returns number of whitespace characters of the
+    same width as if an actual value was supplied.
+
+    Courtesy of 'https://stackoverflow.com/questions/1094841/reusable-library-to-get-human-
+    readable-version-of-file-size'
+
+    :param int size: the size in Bytes
+
+    :return str: the formatted size for output
+    """
+    if size is None:
+        return " " * 10
+    for unit in ["", "Ki", "Mi", "Gi", "Ti"]:
+        if abs(size) < 1024.0:
+            if unit == "":
+                return f"{size:6.0f} B  "
+            return f"{size:6.1f} {unit}B"
+        size /= 1024.0
+    return f"{size:.1f} PiB"
+
+
 class History:
     """Helper with wrapper, which is used when one wants to visualize the version control history
     of the project, printing specific stuff corresponding to a git history
